+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Adds `bijection-json`
 * Adds `bijection-protobuf`
 * Adds `TEnumCodec` for conversion of `TEnum` -> `Int`
+* Change `biject` method in `Bijection` object to `asMethod` for clarity
 
 ### 0.1.0