+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Adds `TEnumCodec` for conversion of `TEnum` -> `Int`
 * Change `biject` method in `Bijection` object to `asMethod` for clarity
 * Adds default `unwrap` bijections to value case classes
+* Adds `Bijection.getOrElse` for conversions between `Option[A]` and `A`
 
 ### 0.1.0